+++ server/src/main/resources/db/migration/V9__change_grade_type.sql
@@ -1,2 +1,2 @@
 ALTER TABLE reviews
-    ALTER COLUMN grade TYPE float8 USING grade::float8;
+    ALTER COLUMN grade TYPE numeric USING grade::numeric;
